pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syn_data_csv"
version = "0.2.0"
description = "Synthetic data generation using LLMs"
readme = "README.md"

authors = [
    { name = "Shyam Kota", email = "shyamsundar.kota@infoservices.com" }
]

maintainers = [
    { name = "Shyam Kota", email = "shyamsundar.kota@infoservices.com" }
]

license = { file = "LICENSE.txt" }

requires-python = ">=3.11"

dependencies = [
    "groq>=0.25.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tqdm>=4.66.4",
    "huggingface_hub>=0.23.0"
]

[project.optional-dependencies]
dev = [
    "twine",
    "build"
]

[project.urls]
Homepage = "https://github.com/Shyamkotainfo/Syn_data_gen/tree/shyam/syn_data_csv"
Documentation = "https://github.com/Shyamkotainfo/Syn_data_gen/blob/shyam/syn_data_csv/README.md"

[tool.setuptools]
packages = ["syn_data_csv", "syn_data_csv.adapters"]
include-package-data = true

[project.scripts]
get_synthetic_data = "syn_data_csv.main:get_csv_data"
